Move some things around, do TOC
Doing on big god commit bc im lazy.  Move some things around, change the
interface of the whole thing to take in a directory and infer the file
locations and such, and also implement table of contents stuff. Note
that we could only do the TOC for djvu files which ultimately are
unlikely to actually work with, but they were the only example we have
that have a TOC, so at least we learned how to mark up the PDF.